pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ops-launcher"
version = "0.1.0"
description = "Interactive terminal Ops Launcher for managing SSH, Docker, and infrastructure across multiple clients and hosts."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{ name="CraftLine Engineering" }]
keywords = ["devops", "ssh", "docker", "terminal", "tui", "ops"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "typer>=0.9,<1.0",
    "rich>=13.0,<14.0",
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.0",
]

[project.scripts]
ops = "ops_launcher.cli:app_entry"

[project.urls]
Homepage = "https://github.com/CraftLineEngineering/ops-launcher"
Repository = "https://github.com/CraftLineEngineering/ops-launcher"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
